Fixed a few bugs
git-svn-id: https://midikit.googlecode.com/svn/trunk@97 c9dcb2c1-a6a4-760b-d948-6d6a4c2e33b0

+++ midi/timer.c
@@ -5,7 +5,7 @@
 #include "message.h"
 
 struct MIDITimer {
-  size_t refs;
+  int refs;
   struct MIDITimerDelegate * delegate;
   MIDILongValue song_position;
   MIDILongValue beats_per_minute;
